Complete Refactoring
